Reorganize nano's UT
