+++ python/nano/test/pytorch/tests/train/test_channels_last.py
@@ -33,7 +33,7 @@ num_classes = 10
 batch_size = 32
 dataset_size = 256
 num_workers = 0
-data_dir = os.path.join(os.path.dirname(__file__), "../data")
+data_dir = "/tmp/data"
 
 
 class CustomResNet(pl.LightningModule):
