+++ python/nano/test/pytorch/tests/train/trainer/test_backprop.py
@@ -36,7 +36,7 @@ from bigdl.nano.pytorch.algorithms.selective_backprop import SelectiveBackprop
 num_classes = 10
 batch_size = 256
 num_workers = 0
-data_dir = os.path.join(os.path.dirname(__file__), "../data")
+data_dir = "/tmp/data"
 
 
 class CheckBatchSize(Callback):
