+++ python/nano/test/pytorch/tests/train/trainer/test_plugin.py
@@ -28,7 +28,7 @@ from bigdl.nano.pytorch import Trainer
 
 from test.pytorch.utils._train_torch_lightning import create_data_loader, data_transform
 from test.pytorch.utils._train_torch_lightning import create_test_data_loader
-from test.pytorch.tests.test_lightning import ResNet18
+from test.pytorch.tests.train.trainer.test_lightning import ResNet18
 
 import copy
 
@@ -36,7 +36,8 @@ num_classes = 10
 batch_size = 32
 dataset_size = 256
 num_workers = 0
-data_dir = os.path.join(os.path.dirname(__file__), "../data")
+data_dir = "/tmp/data"
+
 
 # Linear Model: 1 unit without bias, initialize to 1
 class LinearModel(nn.Module):
@@ -60,8 +61,9 @@ class TestPlugin(TestCase):
 
     def setUp(self):
         test_dir = os.path.dirname(__file__)
+        # project_test_dir = BigDL/python/nano
         project_test_dir = os.path.abspath(
-            os.path.join(os.path.join(os.path.join(test_dir, ".."), ".."), "..")
+            os.path.join(test_dir, "..", "..", "..", "..", "..")
         )
         os.environ['PYTHONPATH'] = project_test_dir
