+++ python/nano/test/pytorch/tests/vision/test_vision_ipex.py
@@ -24,7 +24,7 @@ from test.pytorch.utils._train_torch_lightning import train_with_linear_top_laye
 
 batch_size = 256
 num_workers = 0
-data_dir = os.path.join(os.path.dirname(__file__), "../../data")
+data_dir = "/tmp/data"
 
 
 class TestVisionIPEX(TestCase):
