+++ python/nano/test/run-nano-pytorch-ray-tests.sh
@@ -11,7 +11,7 @@ set -e
 echo "# Start testing"
 start=$(date "+%s")
 python -m pytest -s ${NANO_RAY_TEST_DIR}/test_ray_trainer.py ${NANO_RAY_TEST_DIR}/test_torch_nano_ray.py
-python -m pytest -s ${ANALYTICS_ZOO_ROOT}/python/nano/test/pytorch/tests/test_scale_lr.py -k 'ray'
+python -m pytest -s ${ANALYTICS_ZOO_ROOT}/python/nano/test/pytorch/tests/train/trainer/test_scale_lr.py -k 'ray'
 
 now=$(date "+%s")
 time=$((now-start))
