+++ python/nano/test/run-nano-pytorch-tests-ipex.sh
@@ -4,7 +4,7 @@ export ANALYTICS_ZOO_ROOT=${ANALYTICS_ZOO_ROOT}
 export NANO_HOME=${ANALYTICS_ZOO_ROOT}/python/nano/src
 export PYTORCH_NANO_TEST_DIR=${ANALYTICS_ZOO_ROOT}/python/nano/test/pytorch
 
-wget -nv ${FTP_URI}/analytics-zoo-data/cifar-10-python.tar.gz -P ${PYTORCH_NANO_TEST_DIR}/data
+wget -nv ${FTP_URI}/analytics-zoo-data/cifar-10-python.tar.gz -P /tmp/data
 
 set -e
