+++ python/nano/test/run-nano-pytorch-tests-optimizer.sh
@@ -9,7 +9,7 @@ set -e
 # ipex is not installed here. Any tests needs ipex should be moved to next pytest command.
 echo "# Start testing"
 start=$(date "+%s")
-python -m pytest -s ${PYTORCH_NANO_TEST_DIR}/tests/test_inference_optimizer.py
+python -m pytest -s ${PYTORCH_NANO_TEST_DIR}/tests/inference/test_inference_optimizer.py
 
 now=$(date "+%s")
 time=$((now-start))
